pyproject.toml
[tool.poetry]
name = "lesson-21"
version = "0.1.0"
description = ""
authors = ["qwe qwe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "2.1.0"
selene = "2.0.0rc4"
browserstack-local = "1.2.7"
appium-python-client = "3.1.1"
allure-pytest = "2.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
